requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.1
django-rest-swagger==2.1.2
djangorestframework==3.7.7
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2018.3
requests==2.18.4
simplejson==3.13.2
uritemplate==3.0.0
urllib3==1.22
